% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.ChemoProtSet.R
\docType{methods}
\name{boxplot,ChemoProtSet-method}
\alias{boxplot,ChemoProtSet-method}
\title{Default boxplot for objects of class ChemoProtSet}
\usage{
\S4method{boxplot}{ChemoProtSet}(x, ...)
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{...}{other plotting options}
}
\value{
boxplot for objects of class ChemoProtSet
}
\description{
Description
}
