% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doschedaset.R
\docType{class}
\name{ChemoProtSet-class}
\alias{ChemoProtSet-class}
\alias{DoschedaSet}
\title{An S4 class to run the doscheda pipeline}
\description{
An S4 class to run the doscheda pipeline
}
\section{Slots}{

\describe{
\item{\code{input}}{A data.frame containing the input data}

\item{\code{normData}}{A data.frame containin a processed and standardised version of the input data}

\item{\code{finalData}}{A data.frame containing the final data produced by the pipline}

\item{\code{parameters}}{A list containing all the parameters required to make the pipeline run successfully}

\item{\code{datasets}}{A list containing other potentially useful datasets}
}}

