% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrPlot.R
\docType{methods}
\name{corrPlot}
\alias{corrPlot}
\alias{corrPlot,ChemoProtSet-method}
\alias{corrPlot,ANY,ANY-method}
\title{Plot showing correlation between all channels across replicates}
\usage{
corrPlot(x, ...)

\S4method{corrPlot}{ChemoProtSet}(x, ...)
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{...}{corrplot options}
}
\value{
correlation plot for objects of class ChemoProtSet
}
\description{
Plot of the correlation between all the channels in the data.
}
\examples{

ex <- processedExample
ex <- runNormalisation(ex)
ex <- fitModel(ex)
corrPlot(ex)


}
