% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\docType{methods}
\name{densityPlot}
\alias{densityPlot}
\alias{densityPlot,ChemoProtSet-method}
\alias{densityPlot,ANY,ANY-method}
\title{Density plot for objects of class ChemoProtSet}
\usage{
densityPlot(x, rankProteins = FALSE, ...)

\S4method{densityPlot}{ChemoProtSet}(x, rankProteins = FALSE, ...)
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{rankProteins}{plot a the set of ranked proteins or plot the density of the channels}

\item{...}{other plot options}
}
\value{
density plot for objects of class ChemoProtSet
}
\description{
Description
}
\examples{
ex <- processedExample
ex <- runNormalisation(ex)
ex <- fitModel(ex)
densityPlot(ex)
}
