% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doscheda.R
\docType{package}
\name{doscheda}
\alias{doscheda}
\alias{doscheda-package}
\title{Doscheda: A package for Down Stream Chemo-Proteomics Data Analysis}
\description{
The Doscheda package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

