% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doschedaApp.R
\name{doschedaApp}
\alias{doschedaApp}
\title{Run shiny application for DOSCHEDA}
\usage{
doschedaApp()
}
\value{
Launches shiny application
}
\description{
Run a version of the pipline with some extra features and a simple user experience.
 The application is documented in detail at here
}
