% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{doschedaData}
\alias{doschedaData}
\title{Peptide Intensity data set for Doscheda}
\format{An object of class \code{data.frame} with 21140 rows and 15 columns.}
\usage{
data(doschedaData)
}
\description{
A fabricated data set to run the Doscheda pipeline from peptide intensity.
}
\examples{
data(doschedaData)
head(doschedaData)

}
\keyword{datasets}
