% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{getDatasets}
\alias{getDatasets}
\alias{getDatasets,ChemoProtSet-method}
\alias{getDatasets,ANY,ANY-method}
\title{Accessor function for the datasets slot.}
\usage{
getDatasets(x)

\S4method{getDatasets}{ChemoProtSet}(x)
}
\arguments{
\item{x}{object of class ChemoProtSet}
}
\value{
object of class ChemoProtSet
}
\description{
Accessor function for the datasets slot of a ChemoProtSet object.
}
\examples{
ex <- new('ChemoProtSet')
getDatasets(ex)

}
