% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{getFinal}
\alias{getFinal}
\alias{getFinal,ChemoProtSet-method}
\alias{getFinal,ANY,ANY-method}
\title{Accessor function for the finalData slot.}
\usage{
getFinal(x)

\S4method{getFinal}{ChemoProtSet}(x)
}
\arguments{
\item{x}{object of class ChemoProtSet}
}
\value{
object of class ChemoProtSet
}
\description{
Accessor function for the finalData slot of a ChemoProtSet object.
}
\examples{
ex <- new('ChemoProtSet')
getParameters(ex)

}
