% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{getInput}
\alias{getInput}
\alias{getInput,ChemoProtSet-method}
\alias{getInput,ANY,ANY-method}
\title{Accessor function for the Input}
\usage{
getInput(x)

\S4method{getInput}{ChemoProtSet}(x)
}
\arguments{
\item{x}{object of class ChemoProtSet}
}
\value{
object of class ChemoProtSet
}
\description{
Accessor function for the Input slot of a ChemoProtSet object.
}
\examples{
ex <- new('ChemoProtSet')
getInput(ex)

}
\seealso{
\code{\link{DoschedaSet}}
}
