% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{getNorm}
\alias{getNorm}
\alias{getNorm,ChemoProtSet-method}
\alias{getNorm,ANY,ANY-method}
\title{Accessor function for the normData}
\usage{
getNorm(x)

\S4method{getNorm}{ChemoProtSet}(x)
}
\arguments{
\item{x}{object of class ChemoProtSet}
}
\value{
object of class ChemoProtSet
}
\description{
Accessor function for the normData slot of a ChemoProtSet object.
}
\examples{
ex <- new('ChemoProtSet')
getNorm(ex)

}
