% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{getParameters}
\alias{getParameters}
\alias{getParameters,ChemoProtSet-method}
\alias{getParameters,ANY,ANY-method}
\title{Accessor function for the parameters slot.}
\usage{
getParameters(x)

\S4method{getParameters}{ChemoProtSet}(x)
}
\arguments{
\item{x}{object of class ChemoProtSet}
}
\value{
object of class ChemoProtSet
}
\description{
Accessor function for the parameters slot of a ChemoProtSet object.
}
\examples{
ex <- new('ChemoProtSet')
getParameters(ex)

}
