% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Create report from 'ChemProtSet' object}
\usage{
makeReport(x)
}
\arguments{
\item{x}{Object of class 'ChemoProtSet'}
}
\value{
html report of processed `ChemoProtSet` object
}
\description{
Genereate a report that includes several plots and descriptions for an experiment that has been anaylsed using Doscheda
}
\examples{
\dontrun{
ex<- new('ChemoProtSet')
makeReport(ex)
}
}
