% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSdPlot.R
\docType{methods}
\name{meanSdPlot}
\alias{meanSdPlot}
\alias{meanSdPlot,ChemoProtSet-method}
\alias{meanSdPlot,ANY,ANY-method}
\title{MeanSd plot for objects of class ChemoProtSet}
\usage{
meanSdPlot(x, ...)

\S4method{meanSdPlot}{ChemoProtSet}(x, ...)
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{...}{other plot options}
}
\value{
meanSd plot for objects of class ChemoProtSet
}
\description{
Shows the ranked means with a running median calculated with a window size of 10%
}
\examples{
ex <- processedExample
ex <- runNormalisation(ex)
ex <- fitModel(ex)
meanSdPlot(ex)
}
