% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\docType{methods}
\name{pcaPlot}
\alias{pcaPlot}
\alias{pcaPlot,ChemoProtSet-method}
\alias{pcaPlot,ANY,ANY-method}
\title{PCA of the main data sets contained in a object of class ChemoProtSet}
\usage{
pcaPlot(x, ...)

\S4method{pcaPlot}{ChemoProtSet}(x, ...)
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{...}{other plot options}
}
\value{
PCA plot for objects of class ChemoProtSet
}
\description{
Plot of Principal Component Analysis for the first two principal components of the experimental data.
}
\examples{
ex <- processedExample
ex <- runNormalisation(ex)
ex <- fitModel(ex)
pcaPlot(ex)
ex <- processedExample
ex <- runNormalisation(ex)
ex <- fitModel(ex)
pcaPlot(ex)

}
