% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ChemoProtSet.R
\name{plot.ChemoProtSet}
\alias{plot.ChemoProtSet}
\title{Default plot for objects of class ChemoProtSet}
\usage{
\method{plot}{ChemoProtSet}(x, sigmoidCoef = "rb50", ...)
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{sigmoidCoef}{the sigmoidal coeffcient, one of ('difference', 'slope', 'rb50'). Obselete if modelType is 'linear'}

\item{...}{other plotting options}
}
\value{
plot for objects of class ChemoProtSet
}
\description{
Description
}
