% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{processedExample}
\alias{processedExample}
\title{Processed Peptide Intensity data set for Doscheda}
\format{An object of class \code{ChemoProtSet} of length 1.}
\usage{
data(processedExample)
}
\description{
A processed fabricated data set to run the Doscheda pipeline from peptide intensity.
}
\examples{
data(processedExample)
str(processedExample)
}
\keyword{datasets}
