% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNormalisation.R
\docType{methods}
\name{runNormalisation}
\alias{runNormalisation}
\alias{runNormalisation,ChemoProtSet-method}
\alias{runNormalisation,ANY,ANY-method}
\title{Method to remove peptides from input data of an object of class 'ChemoProtSet'}
\usage{
runNormalisation(x, normalise = "loess")

\S4method{runNormalisation}{ChemoProtSet}(x, normalise = "loess")
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{normalise}{string indicating the type of normalisation that should take place ('loess', 'median', 'none')}
}
\value{
object of class ChemoProtSet
}
\description{
Method to remove peptides from input data of an object of class 'ChemoProtSet'
}
\examples{
ex <- processedExample
ex <- runNormalisation(ex)
ex

}
