% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\docType{methods}
\name{volcanoPlot}
\alias{volcanoPlot}
\alias{volcanoPlot,ChemoProtSet-method}
\alias{volcanoPlot,ANY,ANY-method}
\title{Volcano plot for objects of class ChemoProtSet}
\usage{
volcanoPlot(x, coefficient = "slope", avExprs = 0.2, pVal = 0.05, ...)

\S4method{volcanoPlot}{ChemoProtSet}(x, coefficient = "slope",
  avExprs = 0.2, pVal = 0.05, ...)
}
\arguments{
\item{x}{object of class 'ChemoProtSet'}

\item{coefficient}{coefficient of linear model to be plotted ('slope','intercept','quadratic')}

\item{avExprs}{average expression cutoff}

\item{pVal}{p-value cut-off}

\item{...}{other plotting options}
}
\value{
volcano plot for objects of class ChemoProtSet
}
\description{
Volcano plots designed to be run on objects of class `ChemoProtSet` when a linear model has been applied.
}
\examples{
ex <- processedExample
ex <- runNormalisation(ex)
ex <- fitModel(ex)
volcanoPlot(ex)

}
\seealso{
\code{\link{DoschedaSet}}
}
